scraper/requirements.txt
# Playwright Scraper Dependencies
playwright>=1.40.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.5.0
